keep instfs_uuid when overwriting file with itself
to prevent `replacement_file_id` updates from getting out of hand
when a file gets repeatedly overwritten by itself, Canvas re-uses
the Attachment object in these cases. currently, however, the
object gets the new instfs_uuid and the old one gets deleted
(after checking whether it is in use).

this ends up breaking links to things like canvadocs and
image thumbnails, so approach this from the other side:
leave the instfs_uuid alone in this case, and delete the
new uuid instead. since it never hits the canvas db,
we know it's not used by canvas.

test plan:
 - in an inst-fs configuration,
  - upload an image file
  - re-upload the same image file, overwriting
    the previous copy
  - the image thumbnail shouldn't break

flag=none
refs FOO-5440
refs RCX-3544

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/378567
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>

+++ app/models/attachment.rb
@@ -1843,25 +1843,13 @@ class Attachment < ActiveRecord::Base
     end
   end
 
-  # after replacing this file's instfs_uuid, delete the old file if it's not being used by any other Attachments
-  def safe_delete_overwritten_instfs_uuid(instfs_uuid)
-    raise ArgumentError, "instfs_uuid not overwritten" if self.instfs_uuid == instfs_uuid
-
-    if cloned_item_id.present?
-      # we can't delete the old file since it's used by other attachments;
-      # however, this one isn't using it anymore so we should detach from the clone group
-      update!(cloned_item_id: nil)
-      return
-    end
+  # if we reused this Attachment when an identical file was re-uploaded, delete the unused instfs_uuid
+  def safe_delete_unused_instfs_uuid(instfs_uuid)
+    raise ArgumentError, "instfs_uuid in use" if self.instfs_uuid == instfs_uuid
 
     shard.activate do
-      # any linked Canvadoc retains the old instfs_uuid, and deleting it would break re-rendering
-      return if Canvadoc.where(attachment_id: self).exists?
-
       # double-check that no other attachments are using this instfs_uuid
-      return if Attachment.where(instfs_uuid:).exists?
-
-      InstFS.delete_file(instfs_uuid)
+      InstFS.delete_file(instfs_uuid) unless Attachment.where(instfs_uuid:).exists?
     end
   end
